feat: adds prompt for overwriting when scaffold is rerun
Currently running `scaffold init` a second time results in an error
message being propogated to the user as the cli output. This is
harmless, as the error regards the files being created already existing,
but not a great UX.

This adds promptui to run an interactive prompt for overwriting the
files. If the user selects to overwite, the command behaves as though it
were run with the `-force` option. If the user doesn't, it ends the
command but doesn't return an error.

Other related changes introduced by this change:

 * The cli app is wrapped in a new VervetApp struct, allowing config
   options specific to vervet but not to the CLI interface to be set--
   one of these options is a prompting interface.
 * The vervet app is added to the context passed into the cli app, so
   that vervet specific info can be used in the commands called by the
   cli app.

Fixes #56