fix: prevent empty version folders from being created
The current method to creating the version spec files involved looping over all identified versions and then over the stabilities and creating folders for all, regardless of whether a spec file was necessary.

While not a problem, per se, it conflicted with the mapping of those versions within Registry when looping over the compiled versions and checking for their existence. In that respect, if a folder is found with no file, then that should be a legitimate error.

The solution, simply, is to only create folders for which spec files exist.