refactor: distinction between lint and compile source files
In order to support differential linting against Git repository
branches, linters need to be able to determine which files should be
compared by looking at both the "from" and "to" file sources. This
change introduces support for this, while also making the necessary
distinction between collections of files that need to be compared for
linting, and files that need to be compiled into aggregate OpenAPI
specs.