fix: fix gofmt and other linting on generated embed.FS
This fixes gofmt and adds a godoc comment to the generated embed.go
which embeds the compiled OpenAPI specs for compilation into a Go
service.

Added a test to assert that the godoc comment does not interfere with
the //go:embed magic comments.

Added golangci-lint to the versionware examples and cleaned a few things
up.

Fixes #120.