fix: validate the server URL to prevent subtle parsing effects
Go's net/url.Parse is very permissive. It's easy to pass the same
"host:port" used to configure the listener as the server URL, which gets
mis-interpreted as host://port, when http://host:port was intended.

This adds strict validation to the server URL, preventing this kind of
subtle bug from creating much confusion when the validator then fails to
match any paths and everything 404s.