feat: optic-ci lint exceptions
Allow making very narrow exceptions for specs that do not pass optic-ci
lint rules as they currently are.

Exceptions are well, exceptional, but might need to be made to unblock
work that can't wait for rules to catch up. Exceptions are made based
on the spec file's relative path in the project, and its sha256. So any
subsequent change in the file will not fall under the exception.