feat: support for multiple duplicate build output paths per API
Why?

Some services have a need to write compiled OpenAPI specs to multiple
destinations. This tends to happen in Typescript services in which we
want to write compiled OpenAPI specs to two places:

- src/**, for commiting into git and ingesting into Backstage
- dist/**, for runtime module access to compiled specs

We could just npm script this in these services, but when a project has
several APIs, this can become fragile -- you now have to keep
file-copying-scripts in sync with Vervet config. Miss something, and what
worked in local dev breaks in deployments.

To simplify this and reduce the error surface, Vervet can instead allow
declaring multiple output paths per API.