feat: generate backstage API model entities for compiled versions
If a project contains a Backstage catalog-info.yaml:
- Create/replace API definitions for Vervet compiled API versions
  See
  https://backstage.io/docs/features/software-catalog/descriptor-format#kind-api
  for background on what these are and contain.
- Keep all other developer-maintained content in the file intact,
  including comments, documents that Vervet doesn't recognize or have
  any business messing with.

This is code generation that cooperates with human-authored content in a
multi-document YAML file. Vervet owns the generated API docs, and only
creates/replaces/removes those to reflect the current state of the
service's API versions.

Why?

For services versioning their APIs with Vervet, releases will
automatically show up in Backstage. Vervet is a natural place to do
this, since it builds all the API versions of that service and knows
their lifecycle stage.

This PR also introduces some of the policy changes we're going to soon
introduce for experimental stability versions:
- Experimental releases are not intended to be used long term, but
  either promoted or sunset. So,
- Experimental releases are deprecated upon release, and automatically
  considered sunset after 90 days.