feat: exports generate code
To use vervet's generate functions inside other go tools, this exports
the generate function defined in `internal/cmd/generate.go` as a
function that accepts a param object and executes the code generator.

To avoid duplication, the generate command is updated to set up the
params from the cli options and then uses the new generate package to do
the actual generation.