fix: backstage API: remove version label, tilde from title
The tilde `~` in version strings seem to cause problems for our
Backstage ingest of API entities, and the `api.snyk.io/version` label is
redundant, (same info is in version-date and version-stability) so it is removed.

Also removing the tilde `~` from the API title as it seems to cause the
rendered title in Backstage to get truncated.