fix: Allow nix package to be build on clean systems
proxyVendor was set which assumes that the go dependencies are available
in the building systems nix store, which is rarely the case. This patch
allows nix to download the dependencies when the requested package is
being built.

We also set the version to the last modified date so this flake will get
a new version number without having to update the version number
manually.