fix: support instantiating a Document with an already loaded URL
This adds a new Document constructor which is needed by
vervet-underground when creating Documents for collation.

The private path and url fields are used to indicate the identity of
documents in errors, necessary for logging problems with that service.

To be followed by a VU change to use the new constructor.