fix: improve logging, fix s3 crash and inappropriate loader reuse
Update vervet to latest version to better log collate errors by
identifying the documents which failed to merge.

Fix a panic in the s3 client that can occur when the client fails to
connect. Found this when trying to get localstack to work.

Fix inappropriate openapi3.Loader reuse among multiple documents. The
loader is very stateful and hangs on to references. This can be the
source of hard to isolate and debug problems in collate. It's possible
(though unknowable at this point) that this may clear up strange
conflicts that shouldn't be occuring in a live deployment with the GCS
backend.

Drive-by: use distroless container base