feat: vervet filter command
Introducing a new subcommand `vervet filter`, which filters OpenAPI
documents. This is primarily useful from the CLI when wanting to
generate code or documentation for a subset of the paths in a large API.
Filtering the input OpenAPI document enables this kind of generated
content to be more concise and focused.

Astute readers may notice an overlap in functionality here with the
RemoveElements function used by VU. A future refactoring should
consolidate RemoveElements into a more generalized Filter that can be
used by both VU and the CLI.