fix: disable lint config parsing when lint disabled
When lint flag is set to false, the lint configuration is still parsed
and vetted. In the case of sweater-comb configured with the original
branch, this causes an issue when the .git directory is not there, for
instance when building in Docker ignoring .git dir, terminating the
command execution.

This change removes totally linters configuration when lint flag is set
to false.