fix: make order of apis in generated catalog-info stable
We need to make sure that we always produce a reproducible output when
generating catalog-info.yaml, it should be a pure function of the vervet
config and open api specs.

We are unmarshalling the vervet config into a go map then iterating over
that. Go randomises the order of values in a map so if a project
declares multiple apis in a vervet config then the order of the
resulting api objects in the catalog-info would be random. The only way
to get stability is to use an external ordered structure, in this case
we can use a string slice of the keys.