fix: update vervet dependencies
Bumps Go dependencies to their latest minor versions.

getkin/kin-openapi is in v0 and introduced breaking changes in v0.113.0
that this changeset deals with: https://github.com/getkin/kin-openapi#v01130

OpenAPI Extensions are now of type map[string]interface{} instead of a
map embedded in a struct. This impacts the way Reflect Walk has to be
performed.

A few openapi3 types have now an `extra` unexported field which needs to
be ignored when comparing struct with go-cmp.

A mergeMap generic utility function has been introduced to simplify
merging components and avoid code repetition.