fix: component extension merge
PR #269 introduces a bug with the upgrade of kin-openapi. Component
extensions are not correctly identified and considered as part of
merging specs. However, a nil pointer exception can occur when no
extensions are specified as part of a spec.

This PR correctly handles this nil pointer and add tests to validate
behaviour.