fix: pick newer versions of endpoints when collating over services
If an endpoint is moved from one service to another, then there is a
chance that we will continue to use the path from the old service even
though the newer one has a more recent version.

It is assumed that all services are independent when doing version
resolution, then all the services are merged together. This is fine if
the services do not have overlapping paths but will cause unexpected
behaviour otherwise.

We keep the first spec for each endpoint we find when collating
versions, which means we need to find the most recent version first. To
do this we can change the order we process specs by using the vervet
version before the scrape timestamp - which will mean more recent
versions will be processed first.