fix: prevent similar components from being overwritten
When collating schemas from multiple services, it is very likely that
components with the same name will be defined in several places. We use
a comparison function to detect when this is the case and inline
components if they have already been declared.

This comparison function however ignored fields meant for documenting
fields. Which is fine when it was written as the only concern was
generating openapi specs inner service to validate requests. However now
we are using the collated specs to generate documentation, this
behaviour is very surprising. Especially when the description is
context sensitive.

This will increase the size of our specs slightly, it should not be that
significant. It will require us to recompile our existing specs to fix
the broken versions.