fix: resolve highest stability from versionware
When compiling specs we always show the highest stability level for a
given version of an endpoint. As such we need to use the same logic when
resolving handlers for dealing with that endpoint.

Before this patch it is possible that a request for a beta version will
hit the handler for the beta endpoint, but then the request would be
validated for the GA version of the same endpoint causing the request to
fail.

This aligns endpoint behaviour with what we publicly document for a
specific version.