fix: prevent specs from being generated when nothing has changed
Previously we were generating specs for every stability level at every
version whenever a new version was introduced. This resulted in a new
experimental version triggering a new ga version, when the ga version
had no differences to the previous ga version.

As most people consume just the ga versions of our api specs, this
behaviour introduced a lot of noise which made it much harder to work
out what had changed between versions.

This patch changes the generation so only stabilities that are on or
lower than the one being modified are regenerated. These specs will have
changes due to how stability resolution happens (will resolve to a
higher stability if available) but it will not generate specs where
nothing has changed.

For example if we has a ga api at 2023-01-01 then a experimental api at
2023-02-01 then before we would have generated:
 - 2023-01-01
 - 2023-01-01~beta
 - 2023-01-01~experimental
 - 2023-02-01
 - 2023-02-01~beta
 - 2023-02-01~experimental
and after this patch we generate:
 - 2023-01-01
 - 2023-01-01~beta
 - 2023-01-01~experimental
 - 2023-02-01~experimental