fix: throw error when conflicting components are found
When we roll up versions and copy over paths from older spec, we
naïvely merge the components defined in those specs. This is a problem
when similar named components are defined in two versions of a spec that
is rolled up.

To fix this problem is a lot of effort which we do not have time for
right now, but detecting it is relatively easy. This patch will fail
building specs if we detect such a case so a user can work around it
manually.