feat: add disk based storage interface
We plan to split the api and the scraper into separate processes, as
such we cannot share memory between the two. This makes the default
"memory" storage unusable so will be removed, instead in its place we
can use this new one which uses the local filesystem as a backing.

The implementation is based on the s3 storage implementation, albeit a
bit simplified.