feat: remove memory storage from vervet underground
We plan to split the vervet underground collator and api into separate
processes to help with startup time and availability. Because of this we
cannot offer a memory backed storage any more as the processes will not
share memory.

The memory storage was really only used for testing, this switches that
to the new disk storage which should operate the same.

BREAKING CHANGE: removed memory storage back end