fix: allow overwriting scrape time in disk storage
We use scrape time as a tie breaker when ordering specs with the same
version and stability. For our current storages we assume the last
modified time of the file (or object) is the scrape time. This poses a
problem in tests which assume a monotonic clock when the files are
written too quickly, eg with a tmpfs file system it is possible to write
files in the same nanosecond.

This wasn't an issue when we were using an in-memory store as we stored
the given scrape time as is in the storage. The test which exercises
this (disk_test:TestCollateVersions) explicitly sets the scrape time
using monotonic operations. We need to respect that if we want to
exercise this disk storage in the same way we were testing the memory
storage.

Fixes disk_test:TestCollateVersions so it passes consistently instead of
randomly failing.