feat: use read only storage interface in api handler
This ensures that no api endpoint can write to storage, which gives us
more safety as we want to separate out the scraping into a separate
process. Only having the api concerned about reading the storage will
mean we are able to set a clear boundary between the scraping and the
api.