fix: always fetch specs from cold storage
The implementations of all the storages assume that CollateVersions
would be called before trying to access them. If it wasn't then they
would not return any specs as the VersionIndex and Version methods were
using an in-memory cache and not hitting the backing store.

This does not work now we aren't collating versions in the same process.
We need to make sure we actually get the versions from storage when they
are requested. This will probably be slower, but we can add a cache
later if needed.