feat: merge vervet-underground into main package
Having vervet underground in a separate package means that it is a chore
to update the version of vervet it uses, as we need to release vervet
and then update the references in vervet underground after. This makes
working on this repo harder than it needs to be.

This merges the packages so every release of vervet will also release
vervet underground.

There shouldn't be anything too surprising in this diff, it is mostly
moving files and updating import paths. The entry points have changed
from api and scraper to vu-api and vu-scraper for clarity.

Direct dependencies have had their versions preserved, however keeping
all the indirect dependency versions consistent would be far too much
effort so this resolves them again - from `go mod tidy`.

READMEs have been concatenated, that might need to be revisited...