fix: bump go package version
We released a breaking change previously, which caused our release
pipeline to tag a v7 but we did not change the version on the go
package. As a result no one could use any of the v7 versions.

This aligns the package version with the git tag so people can install
latest again.