fix: update nix flake
Dependencies have changed since the last time it was updated, so it
would have failed to build. This means that nix users can use the latest
version of this tool.

Additionally update the inputs as it has been a while so might as well.

Once this is merged the tool can be run anywhere with

   nix run github:snyk/vervet

presuming a later patch hasn't updated the dependencies again.