fix: report better error when building future versions
Previously we paniced with the message:

    panic: invalid lifecycle (1)

which is not helpful and doesn't give any indication about what went
wrong or how to fix it.

This new error message should be a lot clearer and help users to realise
their mistake.

I see no reason why we have to panic when we get an unreleased version,
so that has been removed.