fix: merge operations from duplicate paths
We were previously overriding whole paths if they were defined in a
newer version. This means that operations are marked as sunset if they
are not redefined in the newer spec. This is surprising behaviour.

This patch will mean that we merge all operations, so future versions do
not need to redeclare operations on each path for them to remain
current.

Running against this latest version will rewrite already generated
specs.