feat: add new codepath for building specs
This codepath is an attempt to simplify the existing build process. It
is a clean room implementation to avoid adopting any unnecessary steps
from the old flow. Implemented as a new command to make it work
alongside the old command.

So far it only merges paths and not any other part of the spec. This
will need to be worked on before it is suitable, however we can begin
testing specs with this new command by releasing in progress versions.

Eventually this will supplant the old flow when we are confident that it
is feature complete.