feat: roll up lower stabilities in simplebuild
This is a change to how we are handling stabilities. Instead of
generating an api spec for each stability at each version, we are only
going to generate a single spec for each version date. Any endpoints
which do not have a GA version will be carried through to the latest
spec but marked as their original stability.

Any endpoint, once a higher stability is released, cannot then go
backwards in stability. For example a new beta version cannot be
released after a ga version for any specific endpoint. Once an endpoint
has hit ga status, all future versions must also be ga.