feat: include component refs in simplebuild
Refs are an OpenAPI concept where you can define part of a spec then use
a JSON reference
[https://datatracker.ietf.org/doc/html/draft-pbryan-zyp-json-ref-03] to
include that block in another part of the document. These are used
extensively at Snyk so we need to make sure that we include the
referenced section in the compiled specs.

Note that this does not check for conflicts, so components that
reference the same path but in different source documents will overwrite
each other silently.