fix: rename conflicting components in simplebuild
If there are refs from two documents that have the same name but
different definitions then we would previously overwrite the oldest one
discovered with the latter one. This obviously produces wrong results.

This patch renames any refs and the components they point to when it
detects a conflict, however doing so mangles any previously generated
documents as the components are not copied but share references between
the documents. This needs to be addressed in a future version.

This behaviour differs from the older build command as that detects
conflicts but will error out the process and tell the user to fix the
conflicts before running.