fix: prevent mangling documents when renaming components
We had a problem when renaming conflicting components. Since the
component object is a reference inside each document, multiple documents
share the same reference so will get the changes propagated back
whenever a later generated document has a component name change - which
will make the older document invalid.

To avoid massive memory expansion we did not want to deep copy all the
component chains, which is both an expensive and complex operation.

Instead we are outputting documents as they are generated instead of
after they are all written. This way it does not matter that they are
mangled after they are generated since the correct version is already
written to disk. This is not 100% ideal as the documents in memory are
mangled but as long as we don't use them after writing we should be
fine.