fix: change ref rename separator from ~ to __
When we rename refs due to conflicts, we add a separator between the
original name and the dynamic component to reduce the chance of future
collisions. We chose a tilde (~) for this due to that character being
very unlikely to naturally occur.

However our loader fails to parse refs with a tilde in the name, and if
our loader fails then other tools will likely also fail. We should aim
for the most compatibility, so a double underscore will work better for
this purpose.