fix: add x-snyk-api-version to document in simplebuild
Previously this annotation was assumed unnecessary as we moved the
version to the operation level, and as such this annotation is on the
operation instead. However this annotation is used for mapping APIs in
our service catalog, so that process will crash if it is not present.

To preserve this backwards compatibility we can add the document version
and assume it is GA stability.