fix: inject version into nix flake
The version is generated at build time, which was skipped in the nix
package. This means that `vervet -version` would always report
"develop".

The version should be the same as the git tag, hopefully this will be
identified by the current ref but will fall back to the last modified
date.