fix: get document version from filename to fix backstage cmds
Previously we were reading the document version from the top level
x-snyk-api-version extension. This extension does not exist on documents
past the pivot date as versions are done per endpoint. This resulted in
all backstage commands failing. (eg vervet backstage update-catalog)

This information is available in the file path, so instead of poluting
our new documents with this extension we can read it from there instead.
This change is backwards compatable with older specs.