feat: perform two stage release
We are encouraging people to pull this tool using `go run ...` which fetches
the source and builds on the users system. This means that --version would
always return "develop". For this to have a useful version number we need to
bake the version number into the source.

We can do this by releasing slightly differently; instead of tagging the head
of main each time something is merged in we can push a new commit which
contains the version number then tag that commit. We then perform the rest of
the release (docker container, npm package) on that tagged commit instead of
the merged one.

This will allow us to assert that a source build is the same as a binary build
and allow accurate version information.