chore: switch from buildGoModule to gomod2nix in nix package
buildGoModule builds all dependencies into one derivation then calculates the
hash of that, so it is not possible to calculate what the hash should be set to
without rebuilding everything. This is not ideal especially in a CI environment
where we don't have a pre-populated nix store.

The gomod2nix system puts each dependency in its own store derivation, which
also helps with sharing dependencies between applications.