fix: error when there are conflicting names
When there are multiple apis defined (eg internal and public) that have specs
with the same name, version, and stability we will generate the same name in
the catalog-info. Since backstage uses the name as the unique identified this
will result in one shadowing the other and not be available in backstage.