feat: add api name to generated backstage api name
Backstage names need to be unique, there is nothing preventing users from
having the same name across their api specs which would be fine up until we
generate the backstage resources. To prevent us burdening developers with the
chore of renaming them, we can add the api name to the name which is guaranteed
to be unique.