fix: switch to gomod2nix version that supports go 1.23
In go 1.23 vendor/modules.txt is required, the mainline version go gomod2nix
does not output that file when building which makes it fail when building a go
modules targeting go versions >1.23.

This branch is just so there is a viable build target until a fix is merged
into mainline gomod2nix.