fix: load outputs from apis with multiple output paths defined
The vervet config allows both Path and Paths to be defined on the output. This
function assumed that only the former was populated so wouldn't return specs if
only the latter was populated.

This seems like a footgun that will happen again so it would be good to change
the types so this doesn't happen again.