feat: remove HEAD fast path for scraper
Most services are built without support for serving a HEAD request to their
openapi endpoints. Currently we fall back to the normal GET path if the service
returns method not supported.

It would take a lot of effort to update each service to support handling HEAD
requests, for not much benefit. The normal GET path is fast enough without
having this fast path.

Having this path is just making noise in metrics for people logging client
errors.