fix: remove unused components after removing elements
We have a process to remove elements from compiled specs to make sure we are
not leaking any data. However those elements could be referencing components
that we did not remove.

This patch does a final pass after elements have been removed to clean up any
components that are no longer referenced.