feat: upgrade Go & vulnerable dependencies
This commit upgrades `github.com/getkin/kin-openapi` to fix a
vulnerability. The issue with that upgrade is that it's a breaking
change (that's what you get by depending on alpha-releases), so in turn
`github.com/tufin/oasdiff` broke.

`github.com/tufin/oasdiff` recently renamed its module to
`github.com/oasdiff/oasdiff`, so to fix it's breakage with regards to
the openapi-module, it had to be upgraded & renamed as well.

That in turn introduced a breaking change (in its minor version
v1.10.17!!!), so the code had to be adjusted too.

These bumps also required us to upgrade to Go 1.24, and as such also
bumps golangci-lint.

Note that the code-change I _believe_ should behave the same as before,
but given that they've completely refactored that `checker` package they
might have introduced new checks that I missed.